backend/requirements.txt
Flask==3.0.2
Flask-Cors==4.0.0
bcrypt==4.1.2
PyJWT==2.8.1
python-dotenv==1.0.1
requests==2.31.0
Werkzeug==3.0.1
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0